requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
absl-py==2.3.0
    # via
    #   aims-dgs4abm (pyproject.toml)
    #   chex
    #   optax
    #   orbax-checkpoint
    #   orbax-export
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
arviz==0.21.0
    # via aims-dgs4abm (pyproject.toml)
asttokens==3.0.0
    # via stack-data
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
chex==0.1.89
    # via optax
click==8.2.1
    # via wandb
colorama==0.4.6
    # via aims-dgs4abm (pyproject.toml)
comm==0.2.2
    # via ipykernel
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.7
    # via orbax-export
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
etils==1.12.2
    # via
    #   optax
    #   orbax-checkpoint
    #   orbax-export
executing==2.2.0
    # via stack-data
flax==0.10.3
    # via aims-dgs4abm (pyproject.toml)
fonttools==4.58.4
    # via matplotlib
fsspec==2025.5.1
    # via etils
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via wandb
google-api-core==2.25.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.1.1
    # via aims-dgs4abm (pyproject.toml)
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
graphviz==0.21
    # via aims-dgs4abm (pyproject.toml)
h5netcdf==1.6.2
    # via arviz
h5py==3.14.0
    # via h5netcdf
humanize==4.12.3
    # via orbax-checkpoint
hydra-core==1.3.2
    # via aims-dgs4abm (pyproject.toml)
idna==3.10
    # via requests
importlib-resources==6.5.2
    # via etils
ipykernel==6.29.5
    # via aims-dgs4abm (pyproject.toml)
ipython==9.3.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jax==0.5.0
    # via
    #   aims-dgs4abm (pyproject.toml)
    #   chex
    #   flax
    #   numpyro
    #   optax
    #   orbax-checkpoint
    #   orbax-export
jax-cuda12-pjrt==0.5.0
    # via jax-cuda12-plugin
jax-cuda12-plugin==0.5.0
    # via jax
jaxlib==0.5.0
    # via
    #   chex
    #   jax
    #   numpyro
    #   optax
    #   orbax-export
jaxtyping==0.3.2
    # via orbax-export
jedi==0.19.2
    # via ipython
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.8
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
marshmallow==3.26.1
    # via dataclasses-json
matplotlib==3.10.0
    # via
    #   aims-dgs4abm (pyproject.toml)
    #   arviz
    #   seaborn
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.1
    # via
    #   jax
    #   jaxlib
    #   tensorstore
msgpack==1.1.1
    # via
    #   flax
    #   orbax-checkpoint
multipledispatch==1.0.0
    # via numpyro
mypy-extensions==1.1.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via
    #   ipykernel
    #   orbax-checkpoint
numpy==2.3.1
    # via
    #   arviz
    #   chex
    #   contourpy
    #   flax
    #   h5py
    #   jax
    #   jaxlib
    #   matplotlib
    #   ml-dtypes
    #   numpyro
    #   optax
    #   orbax-checkpoint
    #   orbax-export
    #   pandas
    #   scipy
    #   seaborn
    #   tensorstore
    #   treescope
    #   xarray
    #   xarray-einstats
numpyro==0.17.0
    # via aims-dgs4abm (pyproject.toml)
nvidia-cublas-cu12==12.9.1.4
    # via
    #   jax-cuda12-plugin
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
nvidia-cuda-cupti-cu12==12.9.79
    # via jax-cuda12-plugin
nvidia-cuda-nvcc-cu12==12.9.86
    # via jax-cuda12-plugin
nvidia-cuda-runtime-cu12==12.9.79
    # via jax-cuda12-plugin
nvidia-cudnn-cu12==9.10.2.21
    # via jax-cuda12-plugin
nvidia-cufft-cu12==11.4.1.4
    # via jax-cuda12-plugin
nvidia-cusolver-cu12==11.7.5.82
    # via jax-cuda12-plugin
nvidia-cusparse-cu12==12.5.10.65
    # via
    #   jax-cuda12-plugin
    #   nvidia-cusolver-cu12
nvidia-nccl-cu12==2.27.5
    # via jax-cuda12-plugin
nvidia-nvjitlink-cu12==12.9.86
    # via
    #   jax-cuda12-plugin
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
omegaconf==2.3.0
    # via hydra-core
opt-einsum==3.4.0
    # via jax
optax==0.2.4
    # via
    #   aims-dgs4abm (pyproject.toml)
    #   flax
orbax-checkpoint==0.11.4
    # via
    #   aims-dgs4abm (pyproject.toml)
    #   flax
    #   orbax-export
orbax-export==0.0.6
    # via aims-dgs4abm (pyproject.toml)
packaging==25.0
    # via
    #   arviz
    #   h5netcdf
    #   hydra-core
    #   ipykernel
    #   marshmallow
    #   matplotlib
    #   wandb
    #   xarray
pandas==2.3.0
    # via
    #   aims-dgs4abm (pyproject.toml)
    #   arviz
    #   seaborn
    #   xarray
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.2.1
    # via matplotlib
platformdirs==4.3.8
    # via
    #   jupyter-core
    #   wandb
prompt-toolkit==3.0.51
    # via ipython
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   orbax-checkpoint
    #   proto-plus
    #   wandb
psutil==7.0.0
    # via
    #   ipykernel
    #   wandb
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via wandb
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   rich
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   flax
    #   omegaconf
    #   orbax-checkpoint
    #   wandb
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
requests==2.32.4
    # via
    #   google-api-core
    #   google-cloud-storage
    #   wandb
rich==14.0.0
    # via flax
rsa==4.9.1
    # via google-auth
scipy==1.16.0
    # via
    #   arviz
    #   jax
    #   jaxlib
    #   xarray-einstats
seaborn==0.13.2
    # via aims-dgs4abm (pyproject.toml)
sentry-sdk==2.31.0
    # via wandb
setproctitle==1.3.6
    # via wandb
setuptools==80.9.0
    # via
    #   arviz
    #   chex
simplejson==3.20.1
    # via orbax-checkpoint
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
stack-data==0.6.3
    # via ipython
tensorstore==0.1.75
    # via
    #   flax
    #   orbax-checkpoint
toolz==1.0.0
    # via chex
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via
    #   aims-dgs4abm (pyproject.toml)
    #   numpyro
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
treescope==0.1.9
    # via flax
typing-extensions==4.14.0
    # via
    #   arviz
    #   chex
    #   etils
    #   flax
    #   orbax-checkpoint
    #   pydantic
    #   pydantic-core
    #   typing-inspect
    #   typing-inspection
    #   wandb
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
wadler-lindig==0.1.7
    # via jaxtyping
wandb==0.20.1
    # via aims-dgs4abm (pyproject.toml)
wcwidth==0.2.13
    # via prompt-toolkit
xarray==2025.6.1
    # via
    #   arviz
    #   xarray-einstats
xarray-einstats==0.9.1
    # via arviz
zipp==3.23.0
    # via etils
